[Modify] combine two `window.ts`
