+++ src/code/platform/window/common/window.ts
@@ -1,3 +1,7 @@
+import { BrowserWindow } from "electron";
+import { Disposable } from "src/base/common/dispose";
+import { Register } from "src/base/common/event";
+import { UUID } from "src/base/node/uuid";
 import { ICLIArguments } from "src/code/platform/environment/common/argument";
 
 export const enum WindowDisplayMode {
@@ -32,5 +36,23 @@ export interface IOpenWindowOpts {
     
     readonly CLIArgv: ICLIArguments;
 
+}
+
+/**
+ * An interface only for {@link WindowInstance}.
+ */
+export interface IWindowInstance extends Disposable {
+    
+    readonly id: number;
+
+    readonly window: BrowserWindow;
+
+    readonly onDidLoad: Register<void>;
+    
+    readonly onDidClose: Register<void>;
+
+    load(): void;
+
+    close(): void;
 }
